+++ src/spot/rest_api/mod.rs
@@ -444,6 +444,48 @@ impl RestApi {
         self.account_api_client.my_allocations(params).await
     }
 
+    /// Query relevant filters
+    ///
+    /// Retrieves the list of [filters](filters.md) relevant to an account on a given symbol. This is the only endpoint that shows if an account has `MAX_ASSET` filters applied to it.
+    /// Weight: 40
+    ///
+    /// # Arguments
+    ///
+    /// - `params`: [`MyFiltersParams`]
+    ///   The parameters for this operation.
+    ///
+    /// # Returns
+    ///
+    /// [`RestApiResponse<models::MyFiltersResponse>`] on success.
+    ///
+    /// # Errors
+    ///
+    /// This function will return an [`anyhow::Error`] if:
+    /// - the HTTP request fails
+    /// - any parameter is invalid
+    /// - the response cannot be parsed
+    /// - or one of the following occurs:
+    ///   - `RequiredError`
+    ///   - `ConnectorClientError`
+    ///   - `UnauthorizedError`
+    ///   - `ForbiddenError`
+    ///   - `TooManyRequestsError`
+    ///   - `RateLimitBanError`
+    ///   - `ServerError`
+    ///   - `NotFoundError`
+    ///   - `NetworkError`
+    ///   - `BadRequestError`
+    ///
+    ///
+    /// For full API details, see the [Binance API Documentation](https://developers.binance.com/docs/binance-spot-api-docs/rest-api/account-endpoints#query-relevant-filters-user_data).
+    ///
+    pub async fn my_filters(
+        &self,
+        params: MyFiltersParams,
+    ) -> anyhow::Result<RestApiResponse<models::MyFiltersResponse>> {
+        self.account_api_client.my_filters(params).await
+    }
+
     /// Query Prevented Matches
     ///
     /// Displays the list of orders that were expired due to STP.