+++ src/spot/websocket_api/models/all_order_lists_response.rs
@@ -30,7 +30,7 @@ pub struct AllOrderListsResponse {
     #[serde(rename = "result", skip_serializing_if = "Option::is_none")]
     pub result: Option<Vec<models::AllOrderListsResponseResultInner>>,
     #[serde(rename = "rateLimits", skip_serializing_if = "Option::is_none")]
-    pub rate_limits: Option<Vec<models::RateLimitsInner>>,
+    pub rate_limits: Option<Vec<models::RateLimits>>,
 }
 
 impl AllOrderListsResponse {