+++ src/spot/websocket_api/models/order_list_cancel_response.rs
@@ -30,7 +30,7 @@ pub struct OrderListCancelResponse {
     #[serde(rename = "result", skip_serializing_if = "Option::is_none")]
     pub result: Option<Box<models::OrderListCancelResponseResult>>,
     #[serde(rename = "rateLimits", skip_serializing_if = "Option::is_none")]
-    pub rate_limits: Option<Vec<models::RateLimitsInner>>,
+    pub rate_limits: Option<Vec<models::RateLimits>>,
 }
 
 impl OrderListCancelResponse {