+++ src/spot/websocket_api/models/sor_order_place_response.rs
@@ -30,7 +30,7 @@ pub struct SorOrderPlaceResponse {
     #[serde(rename = "result", skip_serializing_if = "Option::is_none")]
     pub result: Option<Vec<models::SorOrderPlaceResponseResultInner>>,
     #[serde(rename = "rateLimits", skip_serializing_if = "Option::is_none")]
-    pub rate_limits: Option<Vec<models::RateLimitsInner>>,
+    pub rate_limits: Option<Vec<models::RateLimits>>,
 }
 
 impl SorOrderPlaceResponse {