+++ src/spot/websocket_api/models/ticker24hr_response1.rs
@@ -30,7 +30,7 @@ pub struct Ticker24hrResponse1 {
     #[serde(rename = "result", skip_serializing_if = "Option::is_none")]
     pub result: Option<Box<models::Ticker24hrResponse1Result>>,
     #[serde(rename = "rateLimits", skip_serializing_if = "Option::is_none")]
-    pub rate_limits: Option<Vec<models::RateLimitsInner>>,
+    pub rate_limits: Option<Vec<models::RateLimits>>,
 }
 
 impl Ticker24hrResponse1 {