+++ src/spot/websocket_api/models/ticker_price_response1.rs
@@ -30,7 +30,7 @@ pub struct TickerPriceResponse1 {
     #[serde(rename = "result", skip_serializing_if = "Option::is_none")]
     pub result: Option<Box<models::TickerPriceResponse1Result>>,
     #[serde(rename = "rateLimits", skip_serializing_if = "Option::is_none")]
-    pub rate_limits: Option<Vec<models::RateLimitsInner>>,
+    pub rate_limits: Option<Vec<models::RateLimits>>,
 }
 
 impl TickerPriceResponse1 {