+++ src/spot/websocket_api/models/ticker_response2.rs
@@ -30,7 +30,7 @@ pub struct TickerResponse2 {
     #[serde(rename = "result", skip_serializing_if = "Option::is_none")]
     pub result: Option<Vec<models::TickerResponse2ResultInner>>,
     #[serde(rename = "rateLimits", skip_serializing_if = "Option::is_none")]
-    pub rate_limits: Option<Vec<models::RateLimitsInner>>,
+    pub rate_limits: Option<Vec<models::RateLimits>>,
 }
 
 impl TickerResponse2 {