+++ src/spot/websocket_api/models/user_data_stream_start_response.rs
@@ -30,7 +30,7 @@ pub struct UserDataStreamStartResponse {
     #[serde(rename = "result", skip_serializing_if = "Option::is_none")]
     pub result: Option<Box<models::UserDataStreamStartResponseResult>>,
     #[serde(rename = "rateLimits", skip_serializing_if = "Option::is_none")]
-    pub rate_limits: Option<Vec<models::RateLimitsInner>>,
+    pub rate_limits: Option<Vec<models::RateLimits>>,
 }
 
 impl UserDataStreamStartResponse {