+++ src/spot/websocket_api/models/user_data_stream_stop_response.rs
@@ -30,7 +30,7 @@ pub struct UserDataStreamStopResponse {
     #[serde(rename = "response", skip_serializing_if = "Option::is_none")]
     pub response: Option<serde_json::Value>,
     #[serde(rename = "rateLimits", skip_serializing_if = "Option::is_none")]
-    pub rate_limits: Option<Vec<models::RateLimitsInner>>,
+    pub rate_limits: Option<Vec<models::RateLimits>>,
 }
 
 impl UserDataStreamStopResponse {